+++ crates/hyperqueue/src/client/status.rs
@@ -27,7 +27,7 @@ pub fn job_status(info: &JobInfo) -> Status {
         Status::Canceled
     } else {
         assert_eq!(info.counters.n_finished_tasks, info.n_tasks);
-        if info.is_open {
+        if info.is_open() {
             Status::Opened
         } else {
             Status::Finished